crm/requirements.txt
asgiref==3.8.1
coverage==7.7.1
Django==5.1.7
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.0
iniconfig==2.1.0
packaging==24.2
pluggy==1.5.0
psycopg2==2.9.10
PyJWT==2.9.0
pytest==8.3.5
pytest-cov==6.0.0
sqlparse==0.5.3
